pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "capymind-session"
version = "0.1.0"
description = "Pocket therapist AI agent powered by Google ADK with Gemini fallback"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "CapyMind" }
]
license = { text = "MIT" }
keywords = ["agent", "google-adk", "gemini", "mental health", "cli"]
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.1",
  "google-generativeai>=0.8.0",
]

[project.optional-dependencies]
# Install ADK via: pip install .[adk]
adk = [
  "adk-python @ git+https://github.com/google/adk-python@main",
]

[project.scripts]
capymind-session = "capymind_session.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["capymind_session*"]
